move zenhack.net/go/util into util subdirectory
