+++ util/README.md
@@ -1,0 +1,2 @@
+This directory contains misc. Go utility packages that Ian Denhardt wrote as he
+needed them for other projects.
