+++ util/chkfatal.go
