+++ util/chkfatal_test.go
