+++ util/idempotent.go
